extracting methods to GenomicsDBTestUtils
resurrecting the recently deleted GenomicsDBTest utils with new methods
this is helpful for https://github.com/broadinstitute/gatk-protected/issues/1013
